feat(MovieCollection): Movie의 일급콜렉션 구현
